Renamed analysis to calculator. Redefined Calculator class as abstract.
