+++ mdsuite/calculators/einstein_diffusion_coefficients.py
@@ -22,7 +22,7 @@ from tqdm import tqdm
 
 # Import MDSuite packages
 import mdsuite.utils.meta_functions as meta_functions
-from mdsuite.analysis.analysis import Analysis
+from mdsuite.calculators.calculator import Calculator
 
 # Set style preferences, turn off warning, and suppress the duplication of loading bars.
 plt.style.use('bmh')
@@ -30,7 +30,7 @@ tqdm.monitor_interval = 0
 warnings.filterwarnings("ignore")
 
 
-class EinsteinDiffusionCoefficients(Analysis):
+class EinsteinDiffusionCoefficients(Calculator):
     """
     Class for the Einstein diffusion coefficient implementation
